@@ -1,13 +1,16 @@
 package com.dotcms.rest.api.v1.system.permission;
 
 import com.dotcms.contenttype.exception.NotFoundInDbException;
+import com.dotcms.rest.InitDataObject;
 import com.dotcms.rest.ResponseEntityView;
 import com.dotcms.rest.WebResource;
 import com.dotcms.rest.annotation.NoCache;
 import com.dotcms.rest.annotation.SwaggerCompliant;
+import com.dotcms.rest.exception.BadRequestException;
 import com.dotmarketing.beans.Permission;
 import com.dotmarketing.business.APILocator;
 import com.dotmarketing.business.PermissionAPI;
+import com.dotmarketing.business.Permissionable;
 import com.dotmarketing.business.UserAPI;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
@@ -28,9 +31,12 @@ import org.glassfish.jersey.server.JSONP;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
@@ -53,22 +59,36 @@ import java.util.stream.Stream;
 @Tag(name = "Permissions")
 public class PermissionResource {
 
-    private final WebResource      webResource;
+    private final WebResource webResource;
     private final PermissionHelper permissionHelper;
-    private final UserAPI          userAPI;
+    private final UserAPI userAPI;
+    private final PermissionSaveHelper permissionSaveHelper;
 
     public PermissionResource() {
-
-        this(new WebResource(), PermissionHelper.getInstance(), APILocator.getUserAPI());
+        this(new WebResource(),
+             PermissionHelper.getInstance(),
+             APILocator.getUserAPI(),
+             new PermissionSaveHelper());
     }
-    @VisibleForTesting
-    public PermissionResource(final WebResource      webResource,
-                              final PermissionHelper permissionHelper,
-                              final UserAPI          userAPI) {
 
-        this.webResource      = webResource;
-        this.permissionHelper = permissionHelper;
-        this.userAPI          = userAPI;
+    @VisibleForTesting
+    public PermissionResource(final PermissionSaveHelper permissionSaveHelper) {
+        this(new WebResource(),
+             PermissionHelper.getInstance(),
+             APILocator.getUserAPI(),
+             permissionSaveHelper);
+    }
+
+    @VisibleForTesting
+    public PermissionResource(final WebResource          webResource,
+                              final PermissionHelper     permissionHelper,
+                              final UserAPI              userAPI,
+                              final PermissionSaveHelper permissionSaveHelper) {
+
+        this.webResource          = webResource;
+        this.permissionHelper     = permissionHelper;
+        this.userAPI              = userAPI;
+        this.permissionSaveHelper = permissionSaveHelper;
     }
 
     /**
@@ -86,17 +106,17 @@ public class PermissionResource {
         description = "Load a map of permission type indexed by permissionable types and permissions"
     )
     @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", 
+        @ApiResponse(responseCode = "200",
                     description = "Permissions retrieved successfully",
                     content = @Content(mediaType = "application/json",
                                       schema = @Schema(implementation = ResponseEntityPermissionsByTypeView.class))),
-        @ApiResponse(responseCode = "400", 
+        @ApiResponse(responseCode = "400",
                     description = "Bad request - invalid parameters",
                     content = @Content(mediaType = "application/json")),
-        @ApiResponse(responseCode = "401", 
+        @ApiResponse(responseCode = "401",
                     description = "Unauthorized - authentication required",
                     content = @Content(mediaType = "application/json")),
-        @ApiResponse(responseCode = "403", 
+        @ApiResponse(responseCode = "403",
                     description = "Forbidden - insufficient permissions",
                     content = @Content(mediaType = "application/json"))
     })
@@ -268,6 +288,166 @@ public class PermissionResource {
         return Response.ok(new ResponseEntityPermissionGroupByTypeView(permissionsRoleGroupByTypeMap)).build();
     }
 
+    @Operation(
+        summary = "Get permission metadata",
+        description = "Returns available permission levels and scopes that can be assigned to users and roles"
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200",
+                    description = "Permission metadata retrieved successfully",
+                    content = @Content(mediaType = "application/json",
+                                     schema = @Schema(implementation = ResponseEntityPermissionMetadataView.class))),
+        @ApiResponse(responseCode = "401",
+                    description = "Unauthorized - authentication required",
+                    content = @Content(mediaType = "application/json")),
+        @ApiResponse(responseCode = "403",
+                    description = "Forbidden - frontend user attempted access (backend user required)",
+                    content = @Content(mediaType = "application/json"))
+    })
+    @GET
+    @Path("/")
+    @JSONP
+    @NoCache
+    @Produces({MediaType.APPLICATION_JSON})
+    public ResponseEntityPermissionMetadataView getPermissionMetadata(
+            @Parameter(hidden = true) @Context HttpServletRequest request,
+            @Parameter(hidden = true) @Context HttpServletResponse response) {
+
+        Logger.debug(this, () -> "Retrieving permission metadata");
+
+        new WebResource.InitBuilder(webResource)
+                .requiredBackendUser(true)
+                .requestAndResponse(request, response)
+                .rejectWhenNoUser(true)
+                .init();
+
+        final PermissionMetadataView permissionMetadata = PermissionMetadataView.builder()
+            .levels(PermissionUtils.getAvailablePermissionLevels())
+            .scopes(PermissionUtils.getAvailablePermissionScopes())
+            .build();
+
+        Logger.debug(this, () -> "Permission metadata retrieved successfully");
+
+        return new ResponseEntityPermissionMetadataView(permissionMetadata);
+    }
+
+    /**
+     * Updates permissions for a user's individual role on a specific asset.
+     * This endpoint assigns permissions directly to the user's individual role.
+     * If the asset inherits permissions, inheritance will be broken automatically before saving.
+     * Optionally cascade permissions to descendants (removes their individual permissions).
+     *
+     * @param request HTTP servlet request
+     * @param response HTTP servlet response
+     * @param userId User ID or email address
+     * @param assetId Asset ID (host identifier or folder inode)
+     * @param form Permission assignments to save
+     * @return ResponseEntitySaveUserPermissionsView containing the result
+     * @throws DotDataException if data access fails
+     * @throws DotSecurityException if security check fails
+     */
+    @Operation(
+        summary = "Update user permissions on asset",
+        description = "Saves permissions for a user's individual role on a specific asset (host or folder). " +
+                      "This endpoint assigns permissions directly to the user's individual role. " +
+                      "If the asset inherits permissions, inheritance will be broken automatically before saving. " +
+                      "Optionally cascade permissions to descendants (removes their individual permissions). " +
+                      "Only admin users can update permissions."
+    )
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200",
+                    description = "User permissions updated successfully",
+                    content = @Content(mediaType = "application/json",
+                                      schema = @Schema(implementation = ResponseEntitySaveUserPermissionsView.class))),
+        @ApiResponse(responseCode = "400",
+                    description = "Bad request - invalid input (see error message for details)",
+                    content = @Content(mediaType = "application/json")),
+        @ApiResponse(responseCode = "401",
+                    description = "Unauthorized - authentication required",
+                    content = @Content(mediaType = "application/json")),
+        @ApiResponse(responseCode = "403",
+                    description = "Forbidden - admin access required or user lacks EDIT_PERMISSIONS on asset",
+                    content = @Content(mediaType = "application/json")),
+        @ApiResponse(responseCode = "404",
+                    description = "User or asset not found",
+                    content = @Content(mediaType = "application/json"))
+    })
+    @PUT
+    @Path("/user/{userId}/asset/{assetId}")
+    @JSONP
+    @NoCache
+    @Consumes({MediaType.APPLICATION_JSON})
+    @Produces({MediaType.APPLICATION_JSON})
+    public ResponseEntitySaveUserPermissionsView updateUserPermissions(
+            @Parameter(hidden = true) @Context final HttpServletRequest request,
+            @Parameter(hidden = true) @Context final HttpServletResponse response,
+            @Parameter(description = "User ID or email address", required = true, example = "dotcms.org.1")
+            @PathParam("userId") final String userId,
+            @Parameter(description = "Asset identifier (host ID or folder inode)", required = true, example = "48190c8c-42c4-46af-8d1a-0cd5db894797")
+            @PathParam("assetId") final String assetId,
+            @io.swagger.v3.oas.annotations.parameters.RequestBody(
+                description = "Permission updates to apply. Use empty arrays to remove all permissions for a scope.",
+                required = true,
+                content = @Content(
+                    mediaType = "application/json",
+                    schema = @Schema(implementation = SaveUserPermissionsForm.class)
+                )
+            )
+            final SaveUserPermissionsForm form
+    ) throws DotDataException, DotSecurityException {
+
+        Logger.debug(this, () -> "Updating permissions for user: " + userId + " on asset: " + assetId);
+
+        final InitDataObject initData = new WebResource.InitBuilder(webResource)
+                .requiredBackendUser(true)
+                .requestAndResponse(request, response)
+                .rejectWhenNoUser(true)
+                .init();
+
+        final User requestingUser = initData.getUser();
+
+        if (!requestingUser.isAdmin()) {
+            Logger.warn(this, () -> "Non-admin user " + requestingUser.getUserId() +
+                                  " attempted to update permissions");
+            throw new DotSecurityException("Only admin users can update permissions");
+        }
+
+        if (!UtilMethods.isSet(userId)) {
+            throw new BadRequestException("User ID is required");
+        }
+        if (!UtilMethods.isSet(assetId)) {
+            throw new BadRequestException("Asset ID is required");
+        }
+
+        form.checkValid();
+
+        Logger.debug(this, () -> String.format("PUT /permissions/user/%s/asset/%s requested by %s",
+            userId, assetId, requestingUser.getUserId()));
+
+        final User systemUser = APILocator.getUserAPI().getSystemUser();
+        final User targetUser = loadUserByIdOrEmail(userId, systemUser);
+        final Permissionable asset = permissionSaveHelper.resolveAsset(assetId, systemUser);
+
+        // Security check: User must have EDIT_PERMISSIONS on the asset
+        if (!APILocator.getPermissionAPI().doesUserHavePermission(asset, PermissionAPI.PERMISSION_EDIT_PERMISSIONS, requestingUser)) {
+            Logger.warn(this, () -> String.format("User %s does not have EDIT_PERMISSIONS on asset %s",
+                requestingUser.getUserId(), assetId));
+            throw new DotSecurityException("User does not have permission to edit permissions on this asset");
+        }
+
+        final SaveUserPermissionsView saveResult = permissionSaveHelper.saveUserPermissions(
+            targetUser.getUserId(),
+            assetId,
+            form,
+            requestingUser
+        );
+
+        Logger.info(this, () -> String.format("Successfully updated permissions for user %s on asset %s (requested by %s)",
+            targetUser.getUserId(), assetId, requestingUser.getUserId()));
+
+        return new ResponseEntitySaveUserPermissionsView(saveResult);
+    }
+
     private boolean filter(final PermissionAPI.Type permissionType, final Permission permission) {
 
         return null != permissionType?
@@ -279,5 +459,30 @@ public class PermissionResource {
         final PermissionView view = new PermissionView(permission.getId(), permission.getInode(), permission.getRoleId(),
                 PermissionAPI.Type.findById(permission.getPermission()), permission.isBitPermission(), permission.getType());
         return view;
+    }
+
+    /**
+     * Loads a user by ID or email.
+     * First tries to load by user ID, then falls back to email lookup.
+     *
+     * @param userIdOrEmail User ID or email address
+     * @param systemUser System user for API calls
+     * @return The found User
+     * @throws DotDataException if user lookup fails
+     * @throws DotSecurityException if security check fails
+     */
+    private User loadUserByIdOrEmail(final String userIdOrEmail, final User systemUser)
+            throws DotDataException, DotSecurityException {
+
+        try {
+            return this.userAPI.loadUserById(userIdOrEmail);
+        } catch (com.dotmarketing.business.NoSuchUserException e) {
+            try {
+                return this.userAPI.loadByUserByEmail(userIdOrEmail, systemUser, false);
+            } catch (com.dotmarketing.business.NoSuchUserException ex) {
+                Logger.warn(this, String.format("User not found: %s", userIdOrEmail));
+                throw new com.dotcms.rest.exception.NotFoundException("User not found: " + userIdOrEmail);
+            }
+        }
     }
 }
